requirements.txt
numpy
asyncio
aioconsole
pycryptodome